pyproject.toml
[project]
name = "resumate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "pypdf2==3.0.1",
    "python-docx==0.8.11",
    "python-dotenv>=1.0.1",
    "python-fasthtml>=0.12.1",
    "textract==1.6.3",
]
